addget_ib_devices in common uut a and move get_free_port from test util to common util
