+++ superbench/common/utils/__init__.py
@@ -10,11 +10,6 @@ from superbench.common.utils.lazy_import import LazyImport
 nv_helper = LazyImport('superbench.common.utils.nvidia_helper')
 
 __all__ = [
-    'LazyImport',
-    'SuperBenchLogger',
-    'create_sb_output_dir',
-    'get_sb_config',
-    'logger',
-    'nv_helper',
-    'rotate_dir',
+    'LazyImport', 'SuperBenchLogger', 'create_sb_output_dir', 'get_sb_config', 'logger', 'nv_helper', 'rotate_dir',
+    'network'
 ]
