+++ tests/benchmarks/micro_benchmarks/test_sharding_matmul.py
@@ -7,6 +7,7 @@ import tests.benchmarks.utils as utils
 from tests.helper import decorator
 from superbench.benchmarks import BenchmarkRegistry, Platform, Framework, BenchmarkType, ReturnCode
 from superbench.benchmarks.micro_benchmarks.sharding_matmul import ShardingMatmul, ShardingMode
+from superbench.common.utils import network
 
 
 @decorator.cuda_test
@@ -22,7 +23,7 @@ def test_pytorch_sharding_matmul():
 
     assert (BenchmarkRegistry.is_benchmark_context_valid(context))
 
-    utils.setup_simulated_ddp_distributed_env(1, 0, utils.get_free_port())
+    utils.setup_simulated_ddp_distributed_env(1, 0, network.get_free_port())
     benchmark = BenchmarkRegistry.launch_benchmark(context)
 
     # Check basic information.
